Title: 범위 안의 수 2, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240408/범위 안의 수 2/number-in-range-2.cpp
@@ -14,7 +14,7 @@ int main() {
         }
     }
 
-    avg = sum / cnt;
+    avg = (double)sum / cnt;
 
     cout << fixed;
     cout.precision(1);
